+++ src/main/java/agh/ics/oop/Jungle.java
@@ -6,7 +6,7 @@ public class Jungle {
     public final Vector2d lowerLeft;
     public final Vector2d upperRight;
 
-    public Jungle(int width, int height, float jungleRatio) {
+    public Jungle(int width, int height, double jungleRatio) {
         int newHalfWidth = (int) (width * jungleRatio / 2.0);
         int newHalfHeight = (int) (height * jungleRatio / 2.0);
         Vector2d center = new Vector2d(width / 2, height / 2);
